Define free variable computation over TermContainerType.
